Fix column overflow bug
When the flex container is in column wrap mode, the height of the container,
if max-height is set, will be from the top of the column to the bottom of the
last element, even if other columns are taller. This causes the taller columns
to overflow, covering up stuff below the flex container.

The solution was to set a fixed height on the flex container. This had not been
previously because the cell borders added 2px of height to each cell height,
making the last one overflow if I set height:7em. Setting box-sizing:border-box
lets .day { height:1em; } and .days { height:7em; } play nicely.